♻️ refactor : textstylepannel로 코드 분리  #48
